Team ladder addition
Added a team ladder to the bottom of the page, with a surrounding container. Changed some of the border radius properties on the header and footer. Added a white margin around the whole page. Changed some of the fonts around the page aswell and the colour of the text on the image to make it easier to read.